Move RecoverLocalChangesetsHandler fully to the cpp file
The exposed interface for this type is just a function, so we can slightly
improve build time/size by not exposing any of the implementation details.
