+++ test/object-store/benchmarks/client_reset.cpp
@@ -144,8 +144,7 @@ struct BenchmarkLocalClientReset : public reset_utils::TestClientReset {
             auto history_local = dynamic_cast<sync::ClientHistory*>(wt_local.get_replication()->_get_history_write());
             std::vector<sync::ClientHistory::LocalChange> local_changes =
                 history_local->get_local_changes(current_local_version.version);
-            _impl::client_reset::RecoverLocalChangesetsHandler handler{wt_remote, frozen_local, logger};
-            handler.process_changesets(local_changes, {}); // throws on error
+            _impl::client_reset::process_recovered_changesets(wt_remote, frozen_local, logger, local_changes);
         }
         _impl::client_reset::transfer_group(wt_remote, wt_local, logger, m_mode == ClientResyncMode::Recover);
         if (m_on_post_reset) {
